Created new child components for SummaryDetail
Moved the Ingredients, Instructions, Nutrition, and Categories out of SummaryDetail into their own components, which are child components to SummaryDetail
