+++ src/components/summaryDetail/Instructions.jsx
@@ -1,0 +1,12 @@
+import { nanoid } from 'nanoid';
+
+export default function Instructions({ instructions }) {
+  // Instructions
+  const instructionEls = instructions.map((instruction) => (<p key={nanoid()}>{instruction.display_text}</p>));
+
+  return (
+    <>
+      {instructionEls}
+    </>
+  );
+}
