+++ src/components/summaryDetail/SummaryDetail.jsx
@@ -1,47 +1,32 @@
-import { nanoid } from 'nanoid';
+import Ingredients from './Ingredients';
+import Nutrition from './Nutrition';
+import Instructions from './Instructions';
+import Categories from './Categories';
 
 export default function SummaryDetail({ recipeDetail }) {
   const {
     name, thumbnail_url: thumbnailUrl, video_url: videoUrl, nutrition, instructions, sections, tags,
   } = recipeDetail;
 
-  // Ingredients
-  const ingredientEls = sections.map((section) => section.components.map((component) => (<p key={nanoid()}>{component.raw_text}</p>)));
-
-  // Nutrition
-  let nutritionEls = null;
-  if (Object.keys(nutrition).length === 0) {
-    return nutritionEls = null;
-  }
-  nutritionEls = Object.entries(nutrition)
-    .filter(([key, value]) => key !== 'updated_at')
-    .map(([key, value]) => (
-      <p key={key}>
-        {key}
-        {': '}
-        {value}
-      </p>
-    ));
-
-  // Instructions
-  const instructionEls = instructions.map((instruction) => (<p key={nanoid()}>{instruction.display_text}</p>));
-
-  // Categories
-  const categoryEls = tags
-    .filter((tag) => tag.type === 'dietary' || tag.type === 'meal' || tag.type === 'difficulty')
-    .map((tag) => (<p key={nanoid()}>{tag.display_name}</p>));
-
   return (
     <>
       <h2>{name}</h2>
       <h3>Ingredients</h3>
-      {ingredientEls}
+      <Ingredients
+        sections={sections}
+      />
       <h3>Nutrition</h3>
-      {nutritionEls}
+      <Nutrition
+        nutrition={nutrition}
+      />
       <h3>Instructions</h3>
-      {instructionEls}
+      <Instructions
+        instructions={instructions}
+      />
       <h4>Categories</h4>
-      {categoryEls}
+      <Categories
+        tags={tags}
+      />
     </>
   );
 }
